backend/pyproject.toml
[tool.poetry]
name = "verseforge"
version = "0.1.0"
description = "多智能体网络小说自动写作系统"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
celery = "^5.3.4"
redis = "^5.0.1"
kafka-python = "^2.0.2"
milvus-client = "^2.3.1"
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3